Update 04_Inches_to_Centimeters.py
With int gives the correct result on the console, but the judge system does not approve it and accepts float, which avoids the sensitivity of the system.

+++ Python_Courses/Python_Basics/Python_PB_2024_06/01_First_Steps_in_Coding_Lab/04_Inches_to_Centimeters.py
@@ -1,5 +1,5 @@
 # read user input
-inches = int(input())
+inches = float(input())
 
 # logic
 cms = inches * 2.54
